P5.11: complete protection settings parity (#6)
Complete Avalonia settings parity for 50/51, 27, 59, 59N, 67P, and 67N; add user-defined curve parameters, portable persisted setting groups, process-bus settings handoff, nine protection cards, regression coverage, and standalone cross-platform packaging validation.

+++ src/Arvrel.Application/Evidence/RelayEvidenceBundle.cs
@@ -26,7 +26,7 @@ public sealed record RelayEvidenceSource(
     string? StreamKey,
     string? StreamId,
     int? AppId,
-    int? SampleCounter);
+    long? SampleCounter);
 
 public sealed record RelayEvidenceSettings(
     string GroupName,
@@ -40,7 +40,7 @@ public sealed record RelayEvidenceMeasurement(
     double Residual,
     double FrequencyHz,
     int SamplesPerCycle,
-    int SampleCounter);
+    long SampleCounter);
 
 public sealed record RelayEvidenceTrust(
     bool AllowsMeasurement,
